+++ eskema/__init__.py
@@ -1,0 +1,4 @@
+from eskema import monkey, type
+
+monkey.activate()
+type.init()