+++ eskema/autopk.py
@@ -1,3 +1,5 @@
+import typing as t
+
 import pandas as pd
 
 PK_CANDIDATES_PRIMARY_PREFIXES = ["id"]
@@ -7,17 +9,17 @@ PK_CANDIDATES_SECONDARY_ENGLISH = ["#", "number", "nr"]
 PK_CANDIDATES_SECONDARY_GERMAN = ["nummer", "no", "kennung", "bezeichner", "identifikator"]
 PK_CANDIDATES_SECONDARY = PK_CANDIDATES_SECONDARY_ENGLISH + PK_CANDIDATES_SECONDARY_GERMAN
 
+PEEK_LINES = 1000
 
-def infer_pk(data):
+
+def infer_pk(data: t.Any) -> t.Optional[str]:
     """
-    Attempt to infer primary key from column names and data.
-
-    Aims to implement the algorithm to DWIM [1].
+    Attempt to infer primary key from column names and data, DWIM [1].
 
     [1] https://en.wikipedia.org/wiki/DWIM
     """
-    # Look at the first 1000 lines worth of data.
-    df = pd.read_json(data, lines=True, nrows=1000)
+    # Only peek at the first lines of data.
+    df = pd.read_json(data, lines=True, nrows=PEEK_LINES)
     columns = list(df.columns)
 
     # If there is any column starting with "id", use it as primary key right away.