+++ eskema/cli.py
@@ -5,8 +5,9 @@ from pathlib import Path
 
 import click
 
-from eskema.core import generate_ddl
-from eskema.util import boot_click, sql_canonicalize, sql_pretty
+from eskema.core import SchemaGenerator
+from eskema.model import Resource, SqlTarget
+from eskema.util import boot_click
 
 logger = logging.getLogger(__name__)
 
@@ -21,26 +22,27 @@ def cli(ctx: click.Context, verbose: bool, debug: bool):
 
 
 @cli.command("infer-ddl")
-@click.option("--dialect", type=str, required=True)
+@click.argument("input", type=str, required=True)
+@click.option("--content-type", type=str, required=False)
+@click.option("--dialect", type=str, required=False)
 @click.option("--table-name", type=str, required=False)
 @click.option("--primary-key", type=str, required=False)
-@click.argument("input", type=str, required=True)
 @click.pass_context
 def infer_ddl(
     ctx: click.Context,
-    dialect: str,
+    input: t.Optional[t.Union[Path, str]] = None,  # noqa: A002
+    content_type: t.Optional[str] = None,
+    dialect: t.Optional[str] = None,
     table_name: t.Optional[str] = None,
     primary_key: t.Optional[str] = None,
-    input: t.Optional[t.Union[Path, str]] = None,  # noqa: A002
 ):
     """
     Infer SQL DDL from input data.
     """
-    indata = input
+    indata: t.Union[t.IO, Path, str, None] = input
     path: t.Optional[Path] = None
 
     # Read data from stdin.
-    # TODO: Can also leave empty?
     if indata == "-":
         logger.info("Loading data from stdin")
         if not sys.stdin.readable():
@@ -49,19 +51,28 @@ def infer_ddl(
         indata = sys.stdin
 
     # Read data from file.
-    else:
+    elif isinstance(indata, (Path, str)):
         path = Path(indata)
+        logger.info(f"Loading data from '{path}'")
         indata = path.open("r")
-        logger.info(f"Loading data from {indata}")
 
-    # Derive table name from input file name or data.
-    if table_name is None and path is not None:
-        table_name = path.stem
+    else:
+        raise click.BadArgumentUsage("Unable to read data. 'input' argument missing or empty")
 
     # Decode data.
-    sql = generate_ddl(indata, table_name=table_name, primary_key=primary_key)
-    sql = sql_canonicalize(sql)
-    print(sql_pretty(sql))  # noqa: T201
-    logger.info("Ready.")
+    sg = SchemaGenerator(
+        resource=Resource(
+            data=indata,
+            path=path,
+            content_type=content_type,
+        ),
+        target=SqlTarget(
+            dialect=dialect,
+            table_name=table_name,
+            primary_key=primary_key,
+        ),
+    )
+    result = sg.to_sql_ddl()
+    print(result.pretty)  # noqa: T201
 
     indata.close()