+++ eskema/core.py
@@ -1,57 +1,60 @@
-import io
 import logging
-import typing as t
-from pathlib import Path
 
-from eskema import monkey
 from eskema.autopk import infer_pk
+from eskema.model import Resource, SqlResult, SqlTarget
+from eskema.util import get_firstline
 
-monkey.activate()
+PEEK_BYTES = 10000
 
-from ddlgenerator.ddlgenerator import Table
-
-BYTES_MAGIC = 1000
+logger = logging.getLogger(__name__)
 
 
-def generate_ddl(indata, table_name=None, primary_key=None):
+class SchemaGenerator:
     """
-    Prints code (SQL, SQLAlchemy, etc.) to define a table.
+    Generate SQL DDL schema from input data.
     """
 
-    indata = indata.read(10000)
+    def __init__(
+        self,
+        resource: Resource,
+        target: SqlTarget,
+    ):
+        self.resource = resource
+        self.target = target
+        self.configure()
 
-    # When primary key is not given, try to infer it.
-    if primary_key is None:
-        primary_key = infer_pk(indata)
+    def configure(self):
+        # Sanity checks.
+        if not self.target.dialect:
+            raise ValueError("Inferring the database schema needs an SQLAlchemy dialect")
 
-    firstline = get_firstline(indata)
-    table = Table(
-        firstline,
-        table_name=table_name,
-        varying_length_text=True,
-        uniques=False,
-        pk_name=primary_key,
-        force_pk=True,
-        reorder=False,
-        loglevel=logging.DEBUG,
-        limit=None,
-    )
-    return table.sql(dialect="crate", creates=True, drops=False, inserts=False)
+        # Derive table name from input file name or data.
+        if not self.target.table_name and self.resource.path:
+            self.target.table_name = self.resource.path.stem
 
+        self.resource.configure()
 
-def get_firstline(data: t.Union[Path, str, t.IO]):
-    if isinstance(data, io.TextIOBase):
-        return stream_get_firstline(data)
-    elif isinstance(data, Path):
-        data = Path(data)
-        with open(data, "r") as f:
-            return stream_get_firstline(f)
-    elif isinstance(data, str):
-        return io.StringIO(data.splitlines()[0])
-    else:
-        raise TypeError(f"Unable to decode identification chunk from type {type(data)}")
+    def to_sql_ddl(self) -> SqlResult:
+        from ddlgenerator.ddlgenerator import Table
 
+        # Only peek at the first bytes of data.
+        indata = self.resource.data.read(PEEK_BYTES)
 
-def stream_get_firstline(stream: t.Union[t.IO, str]):
-    firstline = stream.readline()
-    return io.StringIO(firstline)
+        # When primary key is not given, try to infer it from the data.
+        if self.target.primary_key is None:
+            self.target.primary_key = infer_pk(indata)
+
+        firstline = get_firstline(indata)
+        table = Table(
+            firstline,
+            table_name=self.target.table_name,
+            varying_length_text=True,
+            uniques=False,
+            pk_name=self.target.primary_key,
+            force_pk=True,
+            reorder=False,
+            loglevel=logging.DEBUG,
+            limit=None,
+        )
+        sql = table.sql(dialect=self.target.dialect, creates=True, drops=False, inserts=False)
+        return SqlResult(sql)