+++ CHANGES.rst
@@ -13,6 +13,7 @@ in progress
 - Unlock reading from stdin
 - Derive table name from input file name
 - Primary key detection
+- Content type detection using file extension or ``--content-type`` option
 
 
 2023-01-xx 0.0.0